Change password reset and logout HTTP methods to PUT
Changed the HTTP methods for 'resetPassword' and 'logout' endpoints from POST to PUT. This aligns better with REST principles as these methods are technically idempotent - the server's state after the request is made doesn't depend on previous state. Also, added the functionality to revoke the reset token once the password is reset. This increases security by ensuring the token can't be used again.