FindRoomMethod: Fix level ID search & always print room state
Printing this out always helps a lot with debugging when people state that "dive in doesn't work".
Often by the time someone responds "can you take a capture of the API room output", the room state has changed enough where it's not as helpful.
This change makes sure that we can easily see exactly what the server was seeing when it filtered down the rooms.