Make HttpsProxy tell the server which digest to use.
LBP3 gets mad with `/eula` if the server digest is incorrect, and since it doesn't send a client digest for that endpoint, we are unable to guess the correct digest for LBP3 clients.

Also fixes the server locking in the wrong digest on endpoints which expect a server digest, but dont send a client digest. Now it will only lock in the digest for requests which contain a client digest.